Update ActivePanelsContext to store panel ids that are ready to be activated
